json integrantes e outras alerações
