+++ src/views/index.html
@@ -7,7 +7,9 @@
 </head>
 <body>
     <h1>Bem-vindo à nossa aplicação devops da avaliação N1!</h1>
-    <img src=comandos-git.jpeg alt="Tecnologia Git">
+    <img width="500" height="500" src= "comandos-git.jpeg" alt="Tecnologia Git">
     <p>Esta é a página inicial da nossa aplicação, onde vamos desenvolver!</p>
+    
+    <a href="/integrantes"><button>Ver Integrantes</button></a>
 </body>
 </html>
